refactor: patch web3-core
