+++ package.json
@@ -158,7 +158,8 @@
       "@walletconnect/client@1.7.8": "patches/@walletconnect__client@1.7.8.patch",
       "reflect-metadata@0.1.13": "patches/reflect-metadata@0.1.13.patch",
       "bloom-filters@3.0.0": "patches/bloom-filters@3.0.0.patch",
-      "@sentry/core@7.37.2": "patches/@sentry__core@7.37.2.patch"
+      "@sentry/core@7.37.2": "patches/@sentry__core@7.37.2.patch",
+      "web3-core@1.8.0": "patches/web3-core@1.8.0.patch"
     }
   }
 }
