+++ packages/mask/src/plugins/ITO/SNSAdaptor/hooks/useFill.ts
@@ -2,7 +2,7 @@ import { omit } from 'lodash-es'
 import { useAsync, useAsyncFn } from 'react-use'
 import { BigNumber } from 'bignumber.js'
 import { sha3 } from 'web3-utils'
-import type { TransactionReceipt } from 'web3-core/types'
+import type { TransactionReceipt } from 'web3-core'
 import type { ITO2 } from '@masknet/web3-contracts/types/ITO2.js'
 import type { NonPayableTx } from '@masknet/web3-contracts/types/types.js'
 import { TransactionEventType, type ChainId, type SchemaType, FAKE_SIGN_PASSWORD } from '@masknet/web3-shared-evm'
