+++ packages/plugins/EVM/src/state/Connection/middleware/RecentTransaction.ts
@@ -1,4 +1,4 @@
-import type { TransactionReceipt } from 'web3-core/types'
+import type { TransactionReceipt } from 'web3-core'
 import { TransactionStatusType } from '@masknet/web3-shared-base'
 import {
     EthereumMethodType,
