+++ packages/web3-shared/evm/src/types/index.ts
@@ -3,7 +3,7 @@ import type {
     RequestArguments,
     Transaction as Web3Transaction,
     TransactionReceipt as Web3TransactionReceipt,
-} from 'web3-core/types'
+} from 'web3-core'
 import type { JsonRpcPayload, JsonRpcResponse } from 'web3-core-helpers'
 import type { NonPayableTransactionObject, PayableTransactionObject } from '@masknet/web3-contracts/types/types.js'
